code/extractors/hf_extractor/requirements.txt
transformers
datasets
huggingface-hub
numpy
pandas
mysql-connector-python
black
tqdm
kafka-python